Create Mopidy action Invoker
Mopidy and MPD were sharing the same invoker, with Mopidy altering one
of the actions for it's own use.

It turns out, objects returned from `require` functions are cached, so
any alteration to it effects the others in the system. I probably should
have realised that. Bad Dan.

My solution is to seperate the actions object into its own file and have
the invoker duplicate the object when it is created.